made star rating separate component
