+++ ThirdPlace-UI/src/components/submission/AdditionalUserReviews.jsx
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
 import RenderDateAndTime from '../condensed-submission/DateTimeStamp';
+import StarRating from './StarRating';
 
 export default function AdditionalUserReviews({ submissionId }) {
   const [reviews, setReviews] = useState([]);
@@ -31,38 +32,22 @@ export default function AdditionalUserReviews({ submissionId }) {
       });
   }, [submissionId]);
 
-  // Function to render stars based on rating
-  const renderStars = (rating) => {
-
-    const fullStars = Math.floor(rating);
-    const halfStar = rating % 1 !== 0;
-    const stars = [];
-
-    for (let i = 0; i < fullStars; i++) {
-      stars.push("⭐");
-    }
-    if (halfStar) {
-      stars.push("⭐");
-    }
-
-    return stars.join("");
-  };
 
   return (
     <div>
-      <h3>Additional User Reviews</h3>
+      {/* <h3>Additional User Reviews</h3>
       {error && <p>{error}</p>}
       {averageRating !== null && (
         <div>
-          <h4>Average Rating: {renderStars(averageRating) + " " + averageRating}</h4>
+          <h4>Average Rating: <StarRating rating={averageRating} /></h4>
         </div>
-      )}
+      )} */}
       {reviews.length > 0 ? (
         reviews.map(review => (
           <div key={review.id} className="review-card">
             <h4>{review.user.username}</h4>
             <font size="2">Submitted {RenderDateAndTime(review.submission)}</font><br></br>
-            <p>Rating: {renderStars(review.rating)}</p>
+            <p>Rating: <StarRating rating={review.rating} /></p>
             <p>{review.reviewText}</p>
           </div>
         ))
